backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-filter>=24.0,<26.0
django-cors-headers>=4.4,<5.0
mysqlclient>=2.2,<3.0
